2.0.0-release

+++ controllers/applibraries.js
@@ -7,7 +7,7 @@ export var applibraries = {
 	create: function (request, response) {
 		appsservice.read(request.body.appID)
 			.then(function (results) {
-				console.dir( results );
+				
 				if( results.userID === request.user.userID ){
 					service.create( request.body.appID, request.body.libraryID, request.user.userID )
 					.then( function( _results ){
@@ -17,6 +17,9 @@ export var applibraries = {
 					  console.log( _error );
 					  response.send( JSON.stringify( _error ) );
 					});
+				}else {
+					response.setHeader("Status", "401");
+					response.send("Not authorized");
 				}
 			})
 			.catch(function (error) {
@@ -26,11 +29,24 @@ export var applibraries = {
 	},
 
 	delete: function (request, response) {
-		service.delete(request.params.appID, request.params.libraryID, request.user.userID)
-			.then(function (success) {
-				console.log( "Deleted applibrary." );
-				console.dir( success );
-				response.send(true);
+		appsservice.read(request.body.appID)
+			.then(function (results) {
+				
+				if( results.userID === request.user.userID ){
+					service.delete(request.params.appID, request.params.libraryID, request.user.userID)
+						.then(function (success) {
+							console.log( "Deleted applibrary." );
+							console.dir( success );
+							response.send(true);
+						})
+						.catch(function (error) {
+							console.log(error);
+							response.send(JSON.stringify(error));
+						});
+				}else {
+					response.setHeader("Status", "401");
+					response.send("Not authorized");
+				}
 			})
 			.catch(function (error) {
 				console.log(error);
